requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.13
django-cors-headers==4.4.0
django-easy-audit==1.3.6
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
pip==20.2.3
python-dotenv==1.0.1
setuptools==49.2.1
soupsieve==2.5
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
